# perf: reduce alloc on decoder's component expansion