Escape analysis was deciding to move `componentField` into the heap, which could occur due to the risk of stack overflow caused by repeatedly creating the variable inside the loop.
<img width="1134" alt="Screen Shot 2024-03-14 at 10 49 15" src="https://github.com/muktihari/fit/assets/15964188/8c18f675-6291-4eb9-a0b6-5df764e933d6">

So I decide to move the variable declaration outside the loop so we only have one reusable variable and it's no longer escape to the heap. This reduces significant amount of memory usage and allocation, and also improve the decoding speed:
```js
goos: darwin
goarch: amd64
pkg: github.com/muktihari/fit/decoder
cpu: Intel(R) Core(TM) i5-5257U CPU @ 2.70GHz
         │   old.txt    │            new.txt             │
         │    sec/op    │    sec/op     vs base          │
Decode-4   147.8m ± 16%   130.0m ± 39%  ~ (p=0.089 n=10)

         │   old.txt    │               new.txt               │
         │     B/op     │     B/op      vs base               │
Decode-4   73.49Mi ± 0%   67.39Mi ± 0%  -8.31% (p=0.000 n=10)

         │   old.txt    │               new.txt               │
         │  allocs/op   │  allocs/op   vs base                │
Decode-4   1100.0k ± 0%   900.0k ± 0%  -18.18% (p=0.000 n=10)
```